refactored hbss and vbss tests
