+++ src/test/java/org/cicirello/search/ss/HBSSIntCostsTests.java
@@ -31,129 +31,6 @@ import org.junit.jupiter.api.*;
 public class HBSSIntCostsTests extends SharedTestStochasticSampler {
 
   @Test
-  public void testPrecomputeBiases() {
-    for (int n = 0; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      HeuristicBiasedStochasticSampling<Permutation> ch =
-          new HeuristicBiasedStochasticSampling<Permutation>(h);
-      double[] biases = ch.precomputeBiases(n);
-      double expected = 0;
-      for (int i = 1; i <= n; i++) {
-        expected += 1.0 / i;
-        assertEquals(expected, biases[i - 1], 1E-10);
-      }
-    }
-    for (int n = 0; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      HeuristicBiasedStochasticSampling<Permutation> ch =
-          new HeuristicBiasedStochasticSampling<Permutation>(h, 2.0);
-      double[] biases = ch.precomputeBiases(n);
-      double expected = 0;
-      for (int i = 1; i <= n; i++) {
-        expected += 1.0 / (i * i);
-        assertEquals(expected, biases[i - 1], 1E-10);
-      }
-    }
-    for (int n = 0; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      HeuristicBiasedStochasticSampling<Permutation> ch =
-          new HeuristicBiasedStochasticSampling<Permutation>(h, true);
-      double[] biases = ch.precomputeBiases(n);
-      double expected = 0;
-      for (int i = 1; i <= n; i++) {
-        expected += Math.exp(-i);
-        assertEquals(expected, biases[i - 1], 1E-10);
-      }
-    }
-    HeuristicBiasedStochasticSampling.BiasFunction bias =
-        new HeuristicBiasedStochasticSampling.BiasFunction() {
-          @Override
-          public double bias(int rank) {
-            return 1.0 / (rank * rank);
-          }
-        };
-    for (int n = 0; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      HeuristicBiasedStochasticSampling<Permutation> ch =
-          new HeuristicBiasedStochasticSampling<Permutation>(h, bias);
-      double[] biases = ch.precomputeBiases(n);
-      double expected = 0;
-      for (int i = 1; i <= n; i++) {
-        expected += 1.0 / (i * i);
-        assertEquals(expected, biases[i - 1], 1E-10);
-      }
-    }
-  }
-
-  @Test
-  public void testSelect() {
-    for (int n = 2; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      HeuristicBiasedStochasticSampling<Permutation> ch =
-          new HeuristicBiasedStochasticSampling<Permutation>(h);
-      for (int k = 2; k <= n; k++) {
-        double inc = 1.0 / k;
-        double[] values = new double[n];
-        values[0] = inc;
-        for (int i = 1; i < k; i++) {
-          values[i] = values[i - 1] + inc;
-        }
-        double u = 0.0;
-        for (int i = 0; i < k; i++, u += inc) {
-          assertEquals(i, ch.select(values, k, u));
-        }
-        u = inc / 2;
-        for (int i = 0; i < k; i++, u += inc) {
-          assertEquals(i, ch.select(values, k, u));
-        }
-        u = 1.0 - 1E-10;
-        for (int i = k - 1; i >= 0; i--, u -= inc) {
-          assertEquals(i, ch.select(values, k, u));
-        }
-      }
-    }
-  }
-
-  @Test
-  public void testRandomizedSelect() {
-    for (int n = 0; n < 8; n++) {
-      IntProblem problem = new IntProblem();
-      IntHeuristic h = new IntHeuristic(problem, n);
-      HeuristicBiasedStochasticSampling<Permutation> ch =
-          new HeuristicBiasedStochasticSampling<Permutation>(h);
-      for (int k = 1; k <= n; k++) {
-        double[] values = new double[n];
-        for (int i = 0; i < k; i++) values[i] = 100 - i;
-        for (int i = k; i < n; i++) values[i] = 9999;
-        for (int chosenRank = 1; chosenRank <= k; chosenRank++) {
-          int[] indexes = new int[n];
-          for (int i = 0; i < n; i++) indexes[i] = i;
-          String message = "Decreasing: chosenRank:" + chosenRank + " k:" + k;
-          assertEquals(
-              chosenRank - 1, ch.randomizedSelect(indexes, values, k, chosenRank), message);
-        }
-      }
-      for (int k = 1; k <= n; k++) {
-        double[] values = new double[n];
-        for (int i = 0; i < k; i++) values[i] = 2 + i;
-        for (int i = k; i < n; i++) values[i] = 9999;
-        for (int chosenRank = 1; chosenRank <= k; chosenRank++) {
-          int[] indexes = new int[n];
-          for (int i = 0; i < n; i++) indexes[i] = i;
-          String message = "Increasing: chosenRank:" + chosenRank + " k:" + k;
-          assertEquals(
-              k - chosenRank, ch.randomizedSelect(indexes, values, k, chosenRank), message);
-        }
-      }
-    }
-  }
-
-  @Test
   public void testHeuristicNullIncremental() {
     for (int n = 0; n < 4; n++) {
       IntProblem problem = new IntProblem();
